src/collectors/requirements.txt
kafka-python==2.0.2
requests==2.31.0
feedparser==6.0.10
websocket-client==1.7.0
